move tests to geometry_test fix enum #194
